Drop Storybook from the project
