+++ .dockerignore
@@ -1,4 +1,3 @@
 .git
 cypress
 .github
-.storybook
