+++ .github/workflows/node.js.yml
@@ -30,4 +30,3 @@ jobs:
     - run: npm run lint
     - run: npm run test
     - run: npm run build
-    - run: npm run build-storybook
