--- .storybook/preview-head.html
@@ -1,36 +1,0 @@
-<link rel="stylesheet" href="https://unpkg.com/@patternfly/patternfly@4/patternfly.css" crossorigin />
-<link rel="stylesheet" href="https://unpkg.com/@patternfly/patternfly@4/patternfly-addons.css" crossorigin />
-<!-- <script src="https://unpkg.com/keycloak-js@10.0.1/dist/keycloak.min.js"></script> -->
-
-<script>
-  // window.onload = () => {
-  //   keycloak = new Keycloak({
-  //     url: "http://localhost:8180/auth/",
-  //     realm: "master",
-  //     clientId: 'new'
-  //   });
-  //   keycloak.init({ onLoad: 'check-sso', flow: 'standard', responseMode: 'fragment' }).then((authenticated) => {
-  //     if (!authenticated) {
-  //       keycloak.login({ redirectUri: window.location.href });
-  //     }
-  //   }).catch((error) => {
-  //     console.log(error);
-  //     alert('failed to initialize');
-  //   });
-  // }
-</script>
-
-<style>
-  #root {
-    padding: 3rem;
-  }
-
-  .bg-checkerboard {
-    background-image: linear-gradient(45deg, #808080 25%, transparent 25%),
-      linear-gradient(-45deg, #808080 25%, transparent 25%),
-      linear-gradient(45deg, transparent 75%, #808080 75%),
-      linear-gradient(-45deg, transparent 75%, #808080 75%);
-    background-size: 20px 20px;
-    background-position: 0 0, 0 10px, 10px -10px, -10px 0px;
-  }
-</style>
