+++ src/components/data-loader/DataLoader.test.tsx
@@ -1,7 +1,66 @@
 import { render, waitFor } from "@testing-library/react";
-import React from "react";
-import { MockAdminClient } from "../../stories/MockAdminClient";
+import type KeycloakAdminClient from "keycloak-admin";
+import type { ServerInfoRepresentation } from "keycloak-admin/lib/defs/serverInfoRepesentation";
+import React, { ReactNode } from "react";
+import { HashRouter } from "react-router-dom";
+import { AccessContextProvider } from "../../context/access/Access";
+import { AdminClient } from "../../context/auth/AdminClient";
+import { RealmContext } from "../../context/realm-context/RealmContext";
+import { ServerInfoContext } from "../../context/server-info/ServerInfoProvider";
+import serverInfo from "../../context/server-info/__tests__/mock.json";
+import { WhoAmIContextProvider } from "../../context/whoami/WhoAmI";
+import whoamiMock from "../../context/whoami/__tests__/mock-whoami.json";
 import { DataLoader } from "./DataLoader";
+
+/**
+ * This component provides some mocked default react context so that other components can work in a storybook.
+ * In it's simplest form wrap your component like so:
+ * @example
+ *  <MockAdminClient>
+ *    <SomeComponent />
+ *  </MockAdminClient>
+ * @example <caption>With an endpoint, roles => findOneById</caption>
+ *   <MockAdminClient mock={{ roles: { findOneById: () => mockJson } }}>
+ *     <<SomeComponent />
+ *   </MockAdminClient>
+ * @param props mock endpoints to be mocked
+ */
+export const MockAdminClient = (props: {
+  children: ReactNode;
+  mock?: object;
+}) => {
+  return (
+    <HashRouter>
+      <ServerInfoContext.Provider
+        value={serverInfo as unknown as ServerInfoRepresentation}
+      >
+        <AdminClient.Provider
+          value={
+            {
+              ...props.mock,
+              keycloak: {},
+              whoAmI: { find: () => Promise.resolve(whoamiMock) },
+              setConfig: () => {},
+            } as unknown as KeycloakAdminClient
+          }
+        >
+          <WhoAmIContextProvider>
+            <RealmContext.Provider
+              value={{
+                realm: "master",
+                setRealm: () => {},
+                realms: [],
+                refresh: () => Promise.resolve(),
+              }}
+            >
+              <AccessContextProvider>{props.children}</AccessContextProvider>
+            </RealmContext.Provider>
+          </WhoAmIContextProvider>
+        </AdminClient.Provider>
+      </ServerInfoContext.Provider>
+    </HashRouter>
+  );
+};
 
 describe("DataLoader", () => {
   it("loads the data and renders the result", async () => {
