+++ src/i18n.ts
@@ -19,7 +19,6 @@ import events from "./events/messages";
 import realmSettings from "./realm-settings/messages";
 import realmSettingsHelp from "./realm-settings/help";
 import authentication from "./authentication/messages";
-import storybook from "./stories/messages";
 import userFederation from "./user-federation/messages";
 import userFederationHelp from "./user-federation/help";
 import identityProviders from "./identity-providers/messages";
@@ -52,7 +51,6 @@ const initOptions = {
       ...identityProvidersHelp,
       ...userFederation,
       ...userFederationHelp,
-      ...storybook,
     },
   },
   lng: "en",
