--- src/stories/messages.ts
@@ -1,6 +1,0 @@
-export default {
-  storybook: {
-    helpPlaceholder:
-      "Sometimes you need some help and it's nice when the app does that",
-  },
-};
